Move stalker cmds to stalker module
